server/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
asyncpg==0.29.0
redis[asyncio]==5.0.4
PyNaCl==1.5.0
PyJWT==2.8.0
argon2-cffi==23.1.0
python-dotenv==1.0.1
pydantic==2.7.1
pydantic-settings==2.2.1
httpx==0.27.0
python-json-logger==2.0.7
